+++ src/lib/send-to-api.js
@@ -16,14 +16,14 @@ import { withOptions } from '@clevercloud/client/esm/with-options.js';
  * @param {Number?} cacheDelay
  * @return {function(*=): (any | undefined)}
  */
-export function sendToApi ({ apiConfig, signal, cacheDelay, timeout }) {
+export function sendToApi ({ apiConfig = {}, signal, cacheDelay, timeout }) {
 
   return (requestParams) => {
 
     const cacheParams = { ...apiConfig, ...requestParams };
     return withCache(cacheParams, cacheDelay, () => {
 
-      const { API_HOST, ...tokens } = apiConfig;
+      const { API_HOST = 'https://api.clever-cloud.com', ...tokens } = apiConfig;
       return Promise.resolve(requestParams)
         .then(prefixUrl(API_HOST))
         .then(addOauthHeader(tokens))